week1/Project1/requirements.txt
djangorestframework
Django==2.2.5